✨ feat(codec): modified column type to accept a larger number

+++ database/migrations/2022_06_08_141635_create_codec_audios_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
       $table->id();
 
       $table->string('codec', 16);
-      $table->smallInteger('order')->nullable();
+      $table->integer('order')->nullable();
 
       $table->timestamps();
     });
